tests/wpt/tests/infrastructure/testdriver/bidi/bluetooth/simulate_preconnected_peripheral.https.html

<!DOCTYPE html>
<meta charset="utf-8"/>
<title>TestDriver bidi.bluetooth.simulate_preconnected_peripheral method</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js?feature=bidi"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/bidi-bluetooth-helper.js"></script>

<script>
    promise_setup(async () => {
        await test_driver.bidi.bluetooth.simulate_adapter({
            state: "powered-on"
        });
        await test_driver.bidi.bluetooth.request_device_prompt_updated.subscribe();
    });

    promise_test(async (t) => {
        await test_driver.bidi.bluetooth.simulate_preconnected_peripheral({
            address: DEVICE_ADDRESS,
            name: DEVICE_NAME,
            manufacturerData: [],
            knownServiceUuids: []
        });
        const handle_prompt_promise = selectFirstDeviceOnDevicePromptUpdated();
        const [device] = await Promise.all([requestDeviceWithTrustedClick({
            acceptAllDevices: true
        }), handle_prompt_promise]);
        assert_equals(device.name, DEVICE_NAME);
        await test_driver.bidi.bluetooth.disable_simulation();
    }, "simulate a preconnected peripheral.");
</script>
